chore: restructure file directory for FilterList, and FunctionsList, shared components
