+++ src/flows/pipes/RawFluxEditor/FunctionsList/FunctionsList.tsx
@@ -2,9 +2,9 @@ import React, {FC} from 'react'
 
 import {FLUX_FUNCTIONS} from 'src/shared/constants/fluxFunctions'
 import {FluxToolbarFunction} from 'src/types/shared'
-import Fn from 'src/flows/pipes/RawFluxEditor/FluxInjectionOption'
-import FilterList from 'src/flows/shared/FilterList'
-import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionToolTipContent'
+import Fn from 'src/flows/shared/FilterList/InjectionOption'
+import FilterList from 'src/flows/shared/FilterList/FilterList'
+import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionsList/per-function/FunctionToolTipContent'
 
 interface Props {
   onSelect: (fn: FluxToolbarFunction) => void
